requirements.txt
streamlit
pyodbc
pandas
plotly
scikit-learn
SQLAlchemy
python-dotenv
